refactor(db): rename schema to schema.sqlite.ts with entry point
Move table definitions from schema.ts to schema.sqlite.ts to make room
for a future schema.pg.ts. Create a thin schema.ts entry point that
re-exports from both schema.sqlite.ts and schema.relations.ts. Update
schema.relations.ts to import directly from schema.sqlite.ts to avoid
a circular dependency through schema.ts.
